Replaced VS-Codium with geany
Due to uncertainties related to licensing policies in VSCodium (a community-driven distribution based on VS Code that removes telemetry easily so you don't have to build VScode from source), we are switching to Geany. Geany offers a more independent project framework with established design and licensing considerations that align better with our project's requirements for security, reliability, and freedom.